create new tracking helper and update subscan snowbrigde tracking files
